Fix a few clangtidy warnings.

+++ src/crc16/src/crc16.cpp
@@ -7,8 +7,8 @@ Crc16::Crc16(uint16_t initValue, uint16_t polynomial) :
 
 void Crc16::next(uint8_t value) {
   crc_ ^= ((uint16_t)value << 8);
-  for (int i = 0; i < 8; i++) {
-    if (crc_ & 0x8000) {
+  for(uint8_t i = 0U; i < 8U; i++) {
+    if(crc_ & 0x8000) {
       crc_ = (crc_ << 1) ^ polynomial_;
     }
     else {
@@ -18,7 +18,7 @@ void Crc16::next(uint8_t value) {
 }
 
 void Crc16::next(const uint8_t* values, uint32_t length) {
-  for (uint32_t i = 0; i < length; i++) {
+  for(uint32_t i = 0; i < length; i++) {
     next(values[i]);
   }
 }
